+++ source/render/hwgl/gl_local.h
@@ -398,7 +398,7 @@ public:
   virtual void FillRect (float x1, float y1, float x2, float y2, vuint32 color, float alpha=1.0f) override;
   virtual void DrawRect (float x1, float y1, float x2, float y2, vuint32 color, float alpha=1.0f) override;
   virtual void ShadeRect (float x1, float y1, float x2, float y2, float darkening) override;
-  virtual void DrawLine (float x1, float y1, float x2, float y2, vuint32 color, float alpha=1.0f) override;
+  virtual void DrawLine (int x1, int y1, int x2, int y2, vuint32 color, float alpha=1.0f) override;
   virtual void DrawConsoleBackground (int h) override;
   virtual void DrawSpriteLump (float x1, float y1, float x2, float y2,
                                VTexture *Tex, VTextureTranslation *Translation, bool flip) override;
